% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binosim.R
\name{binosim}
\alias{binosim}
\title{Simulate binary data from logistic regression.}
\usage{
binosim(n, px, baseprob, or)
}
\arguments{
\item{n}{Sample size.}

\item{px}{Probability of x.}

\item{baseprob}{Baseline probability of Y.}

\item{or}{True odds ratio.}
}
\value{
A data frame with binary data
}
\description{
Simulate binary data from logistic regression.
}
\examples{
sim_data <- binosim(n = 100000, px = 0.5, baseprob = 0.3, or = 2)
sim_data
}
\author{
Amrit Tiwana
}
