% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binosim.R
\name{binosim}
\alias{binosim}
\title{Simulate binary data from logistic regression.}
\usage{
binosim(n, px, baselineprob, or)
}
\arguments{
\item{n}{Sample size.}

\item{px}{Probability of x.}

\item{baselineprob}{Baseline probability of Y.}

\item{or}{True odds ratio.}
}
\value{
A data frame with binary data
}
\description{
Simulate binary data from logistic regression.
}
\examples{
sim_data <- binosim(n = 10, px = 0.6, baselineprob = 1, or = 1)
sim_data
}
\author{
Amrit Tiwana
}
